Reduce amount of common code
